Change all styling to be non-conditional
That is, any place where a syle function was before applied only if its
modified item was not of a certain type are now non-conditional and will
always apply. This is to simplify behavior of the `show-*` functions and
make it clear that the place to change styling is in the `style` module.